requirements.txt
http://www.djangoproject.com/download/1.4-alpha-1/tarball/
icalendar==2.2
python-dateutil>=1.5,<2.0
-e hg+http://bitbucket.org/andrewgodwin/south/@64fdcc52cd01#egg=South
python-ldap
django-auth-ldap
URLObject>=2.0.0
-e git+https://github.com/tomchristie/django-rest-framework.git#egg=djangorestframework
